Add support for Turbo vs UJS in remove link
This takes the same approach as in this `administrate` PR [1] by adding
the `data-turbo-method` and `data-turbo-confirm` attributes supported by
Hotwire Turbo [2,3] to the link for removing an attachment.

The existing `method` option and `data-confirm` attribute are left in
place to avoid breaking the behaviour under UJS [4] which was deprecated
in Rails v7 and removed in Rails v7.2.

As the `administrate` PR says, ideally the `link_to` would be replaced
by `button_to`. However, in the interests of making the change as small
as possible and retaining consistency with `administrate` itself, it
seems OK to keep the `link_to` in place.

[1]: https://github.com/thoughtbot/administrate/pull/2448
[2]: https://turbo.hotwired.dev/handbook/drive#performing-visits-with-a-different-method
[3]: https://turbo.hotwired.dev/handbook/drive#requiring-confirmation-for-a-visit
[4]: https://api.rubyonrails.org/v7.0/classes/ActionView/Helpers/UrlHelper.html#method-i-link_to-label-Deprecated-3A+Rails+UJS+Attributes